// src/generated/client/crystalball/military/v1/service_client.ts
export interface GetTheaterPostureResponse {
  theaters: TheaterPosture[];
}

export interface TheaterPosture {
  theater: string;
  postureLevel: string;
  activeFlights: number;
  trackedVessels: number;
  activeOperations: string[];
  assessedAt: number;
}